requirements.txt
python-telegram-bot==20.3
pandas>=1.5.0
matplotlib>=3.5.0
xlsxwriter>=3.0.0
openpyxl>=3.0.0
python-dotenv>=1.0.0